fix(courses): move migrate_edx_data test file out of management/commands
Django's command auto-discovery scans every .py file directly under
management/commands/ as a candidate command module, so
test_migrate_edx_data.py showed up (and would fail) in ./manage.py -h.
Move it to management/tests/, matching the existing convention used by
every other management command test in this app (and in users/).
